services is free from dep on DTO project, work to migrate DTO to API project ready to begin
